change insight auth in docs (#5681)
## Problem solved

https://linear.app/thirdweb/issue/INSIGHT-480/improvements-required-for-insight-playground

Updated the authentication method in the Insight API documentation to use the `x-client-id` header instead of embedding the client ID in the URL path.

<!-- start pr-codex -->

---

## PR-Codex overview
This PR updates the GitHub workflow and documentation related to the `insight` feature. It enhances the valid issue prefixes and provides clearer instructions for authentication in the API usage.

### Detailed summary
- Updated `VALID_ISSUE_PREFIXES` to include `INSIGHT`.
- Improved documentation in `apps/portal/src/app/insight/get-started/page.mdx`:
  - Corrected "clientId" to "client ID".
  - Added a note about using the `x-client-id` header for authentication.
- Modified the API fetch request to include headers for `client ID`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->