[DASH-605] Add profile OG image (#5775)
## Problem solved

Short description of the bug fixed or feature added

<!-- start pr-codex -->

---

## PR-Codex overview
This PR primarily focuses on refactoring the handling of gradients for avatars and updating the Open Graph image generation for user profiles. It introduces a new `GradientBlobbie` component and modifies the `GradientAvatar` to use this new component, enhancing the visual representation of user profiles.

### Detailed summary
- Deleted `url-utils.ts` and `profile.tsx` files.
- Introduced `title` and `description` variables in the profile page response.
- Created `GradientBlobbie` component for gradient backgrounds.
- Removed gradient logic from `GradientAvatar` and used `GradientBlobbie` instead.
- Updated Open Graph image generation to include `GradientBlobbie` for users without an ENS image.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->