Update billing management text (#6258)
<!--

## title your PR with this format: "[SDK/Dashboard/Portal] Feature/Fix: Concise title for the changes"

If you did not copy the branch name from Linear, paste the issue tag here (format is TEAM-0000):

## Notes for the reviewer

Anything important to call out? Be sure to also clarify these in your comments.

## How to test

Unit tests, playground, etc.

-->

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on updating the `PlanInfoCard` component in the billing settings of the dashboard. It enhances the layout and adds informative text to improve user experience.

### Detailed summary
- Changed the inner `div` from `flex items-center` to `flex flex-col items-start`.
- Added a new paragraph with instructions for managing billing.
- Updated the `variant` prop of the `BillingPortalButton` from `outline` to `primary`.

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->