Deprecate pack (#6823)
<!--

## title your PR with this format: "[SDK/Dashboard/Portal] Feature/Fix: Concise title for the changes"

If you did not copy the branch name from Linear, paste the issue tag here (format is TEAM-0000):

## Notes for the reviewer

Anything important to call out? Be sure to also clarify these in your comments.

## How to test

Unit tests, playground, etc.

-->

<!-- start pr-codex -->

---

## PR-Codex overview
This PR focuses on deprecating the `Pack` contract in the `thirdweb` library due to its incompatibility with the Pectra update. Support for this contract will be removed in the next release.

### Detailed summary
- Added deprecation notice for the `Pack` contract in:
  - `packages/thirdweb/src/extensions/pack/createNewPack.ts`
  - `packages/thirdweb/src/extensions/prebuilts/deploy-pack.ts`

> ✨ Ask PR-Codex anything about this PR by commenting with `/codex {your question}`

<!-- end pr-codex -->